+++ build-macosx/build.sh
@@ -19,27 +19,31 @@ source /root/common/prep.sh
 if [ "${CLASSICAL}" == "1" ]; then
   echo "Starting classical build for macOS..."
 
-  $SCONS platform=osx $OPTIONS arch=x86_64 tools=yes target=release_debug
-  $SCONS platform=osx $OPTIONS arch=arm64 tools=yes target=release_debug
-  lipo -create bin/godot.osx.opt.tools.x86_64 bin/godot.osx.opt.tools.arm64 -output bin/godot.osx.opt.tools.universal
-  $STRIP bin/godot.osx.opt.tools.universal
+  if [ "${BUILD_EDITOR_x86_64}" == "1" ] || [ "${BUILD_EDITOR_x86}" == "1" ]; then
+    $SCONS platform=osx $OPTIONS arch=x86_64 tools=yes target=release_debug
+    $SCONS platform=osx $OPTIONS arch=arm64 tools=yes target=release_debug
+    lipo -create bin/godot.osx.opt.tools.x86_64 bin/godot.osx.opt.tools.arm64 -output bin/godot.osx.opt.tools.universal
+    $STRIP bin/godot.osx.opt.tools.universal
 
-  mkdir -p /root/out/tools
-  cp -rvp bin/* /root/out/tools
-  rm -rf bin
+    mkdir -p /root/out/tools
+    cp -rvp bin/* /root/out/tools
+    rm -rf bin
+  fi
 
-  $SCONS platform=osx $OPTIONS arch=x86_64 tools=no target=release_debug
-  $SCONS platform=osx $OPTIONS arch=arm64 tools=no target=release_debug
-  lipo -create bin/godot.osx.opt.debug.x86_64 bin/godot.osx.opt.debug.arm64 -output bin/godot.osx.opt.debug.universal
-  $STRIP bin/godot.osx.opt.debug.universal
-  $SCONS platform=osx $OPTIONS arch=x86_64 tools=no target=release
-  $SCONS platform=osx $OPTIONS arch=arm64 tools=no target=release
-  lipo -create bin/godot.osx.opt.x86_64 bin/godot.osx.opt.arm64 -output bin/godot.osx.opt.universal
-  $STRIP bin/godot.osx.opt.universal
+  if [ "${BUILD_TEMPLATES}" == "1" ]; then
+    $SCONS platform=osx $OPTIONS arch=x86_64 tools=no target=release_debug
+    $SCONS platform=osx $OPTIONS arch=arm64 tools=no target=release_debug
+    lipo -create bin/godot.osx.opt.debug.x86_64 bin/godot.osx.opt.debug.arm64 -output bin/godot.osx.opt.debug.universal
+    $STRIP bin/godot.osx.opt.debug.universal
+    $SCONS platform=osx $OPTIONS arch=x86_64 tools=no target=release
+    $SCONS platform=osx $OPTIONS arch=arm64 tools=no target=release
+    lipo -create bin/godot.osx.opt.x86_64 bin/godot.osx.opt.arm64 -output bin/godot.osx.opt.universal
+    $STRIP bin/godot.osx.opt.universal
 
-  mkdir -p /root/out/templates
-  cp -rvp bin/* /root/out/templates
-  rm -rf bin
+    mkdir -p /root/out/templates
+    cp -rvp bin/* /root/out/templates
+    rm -rf bin
+  fi
 fi
 
 # Mono
@@ -58,40 +62,44 @@ if [ "${MONO}" == "1" ]; then
   # libmono-native-compat.dylib).
   mkdir -p tmp-lib/{x86_64,arm64}
 
-  $SCONS platform=osx $OPTIONS $OPTIONS_MONO mono_prefix=$MONO_PREFIX_X86_64 arch=x86_64 tools=yes target=release_debug copy_mono_root=yes
-  cp bin/GodotSharp/Mono/lib/*.dylib tmp-lib/x86_64/
-  $SCONS platform=osx $OPTIONS $OPTIONS_MONO mono_prefix=$MONO_PREFIX_ARM64 arch=arm64 tools=yes target=release_debug copy_mono_root=yes
-  cp bin/GodotSharp/Mono/lib/*.dylib tmp-lib/arm64/
-  lipo -create bin/godot.osx.opt.tools.x86_64.mono bin/godot.osx.opt.tools.arm64.mono -output bin/godot.osx.opt.tools.universal.mono
-  $STRIP bin/godot.osx.opt.tools.universal.mono
+  if [ "${BUILD_EDITOR_x86_64}" == "1" ] || [ "${BUILD_EDITOR_x86}" == "1" ]; then
+    $SCONS platform=osx $OPTIONS $OPTIONS_MONO mono_prefix=$MONO_PREFIX_X86_64 arch=x86_64 tools=yes target=release_debug copy_mono_root=yes
+    cp bin/GodotSharp/Mono/lib/*.dylib tmp-lib/x86_64/
+    $SCONS platform=osx $OPTIONS $OPTIONS_MONO mono_prefix=$MONO_PREFIX_ARM64 arch=arm64 tools=yes target=release_debug copy_mono_root=yes
+    cp bin/GodotSharp/Mono/lib/*.dylib tmp-lib/arm64/
+    lipo -create bin/godot.osx.opt.tools.x86_64.mono bin/godot.osx.opt.tools.arm64.mono -output bin/godot.osx.opt.tools.universal.mono
+    $STRIP bin/godot.osx.opt.tools.universal.mono
 
-  # Make universal versions of the dylibs we use.
-  lipo -create tmp-lib/x86_64/libmono-native-compat.dylib tmp-lib/arm64/libmono-native.dylib -output tmp-lib/libmono-native.dylib
-  lipo -create tmp-lib/x86_64/libMonoPosixHelper.dylib tmp-lib/arm64/libMonoPosixHelper.dylib -output tmp-lib/libMonoPosixHelper.dylib
-  # Somehow only included in x86_64 build.
-  cp tmp-lib/x86_64/libmono-btls-shared.dylib tmp-lib/
+    # Make universal versions of the dylibs we use.
+    lipo -create tmp-lib/x86_64/libmono-native-compat.dylib tmp-lib/arm64/libmono-native.dylib -output tmp-lib/libmono-native.dylib
+    lipo -create tmp-lib/x86_64/libMonoPosixHelper.dylib tmp-lib/arm64/libMonoPosixHelper.dylib -output tmp-lib/libMonoPosixHelper.dylib
+    # Somehow only included in x86_64 build.
+    cp tmp-lib/x86_64/libmono-btls-shared.dylib tmp-lib/
 
-  cp -f tmp-lib/*.dylib bin/GodotSharp/Mono/lib/
+    cp -f tmp-lib/*.dylib bin/GodotSharp/Mono/lib/
 
-  mkdir -p /root/out/tools-mono
-  cp -rvp bin/* /root/out/tools-mono
-  rm -rf bin
+    mkdir -p /root/out/tools-mono
+    cp -rvp bin/* /root/out/tools-mono
+    rm -rf bin
+  fi
 
-  $SCONS platform=osx $OPTIONS $OPTIONS_MONO mono_prefix=$MONO_PREFIX_X86_64 arch=x86_64 tools=no target=release_debug
-  $SCONS platform=osx $OPTIONS $OPTIONS_MONO mono_prefix=$MONO_PREFIX_ARM64 arch=arm64 tools=no target=release_debug
-  lipo -create bin/godot.osx.opt.debug.x86_64.mono bin/godot.osx.opt.debug.arm64.mono -output bin/godot.osx.opt.debug.universal.mono
-  $STRIP bin/godot.osx.opt.debug.universal.mono
-  $SCONS platform=osx $OPTIONS $OPTIONS_MONO mono_prefix=$MONO_PREFIX_X86_64 arch=x86_64 tools=no target=release
-  $SCONS platform=osx $OPTIONS $OPTIONS_MONO mono_prefix=$MONO_PREFIX_ARM64 arch=arm64 tools=no target=release
-  lipo -create bin/godot.osx.opt.x86_64.mono bin/godot.osx.opt.arm64.mono -output bin/godot.osx.opt.universal.mono
-  $STRIP bin/godot.osx.opt.universal.mono
+  if [ "${BUILD_TEMPLATES}" == "1" ]; then
+    $SCONS platform=osx $OPTIONS $OPTIONS_MONO mono_prefix=$MONO_PREFIX_X86_64 arch=x86_64 tools=no target=release_debug
+    $SCONS platform=osx $OPTIONS $OPTIONS_MONO mono_prefix=$MONO_PREFIX_ARM64 arch=arm64 tools=no target=release_debug
+    lipo -create bin/godot.osx.opt.debug.x86_64.mono bin/godot.osx.opt.debug.arm64.mono -output bin/godot.osx.opt.debug.universal.mono
+    $STRIP bin/godot.osx.opt.debug.universal.mono
+    $SCONS platform=osx $OPTIONS $OPTIONS_MONO mono_prefix=$MONO_PREFIX_X86_64 arch=x86_64 tools=no target=release
+    $SCONS platform=osx $OPTIONS $OPTIONS_MONO mono_prefix=$MONO_PREFIX_ARM64 arch=arm64 tools=no target=release
+    lipo -create bin/godot.osx.opt.x86_64.mono bin/godot.osx.opt.arm64.mono -output bin/godot.osx.opt.universal.mono
+    $STRIP bin/godot.osx.opt.universal.mono
 
-  cp -f tmp-lib/*.dylib bin/data.mono.osx.64.release/Mono/lib/
-  cp -f tmp-lib/*.dylib bin/data.mono.osx.64.release_debug/Mono/lib/
+    cp -f tmp-lib/*.dylib bin/data.mono.osx.64.release/Mono/lib/
+    cp -f tmp-lib/*.dylib bin/data.mono.osx.64.release_debug/Mono/lib/
 
-  mkdir -p /root/out/templates-mono
-  cp -rvp bin/* /root/out/templates-mono
-  rm -rf bin
+    mkdir -p /root/out/templates-mono
+    cp -rvp bin/* /root/out/templates-mono
+    rm -rf bin
+  fi
 fi
 
 echo "macOS build successful"
